Start with integration test for distributeTAs
