+++ src/test/java/nl/tudelft/ewi/devhub/webtests/AssignmentTest.java
@@ -11,6 +11,8 @@ import nl.tudelft.ewi.devhub.server.database.entities.Group;
 import nl.tudelft.ewi.devhub.webtests.utils.WebTest;
 import nl.tudelft.ewi.devhub.webtests.views.AssignmentView;
 import nl.tudelft.ewi.devhub.webtests.views.AssignmentsView;
+import nl.tudelft.ewi.devhub.webtests.views.CourseView;
+import nl.tudelft.ewi.devhub.webtests.views.LoginView;
 import org.junit.Test;
 
 import java.text.ParseException;
@@ -21,7 +23,7 @@ import static org.junit.Assert.assertEquals;
 public class AssignmentTest extends WebTest {
 
     @Inject
-	private CourseEditions courseEditions;
+    private CourseEditions courseEditions;
 
     @Inject
     private Deliveries deliveries;
@@ -29,39 +31,39 @@ public class AssignmentTest extends WebTest {
     @Inject
     private Groups groups;
 
-	@Test
-	public void testListAssignments() {
-		AssignmentsView view = openLoginScreen()
-				.login(NET_ID, PASSWORD)
-				.toCoursesView()
-				.listMyProjects()
-				.get(0).click()
-				.toAssignmentView();
-		
+    @Test
+    public void testListAssignments() {
+        AssignmentsView view = openLoginScreen()
+                .login(NET_ID, PASSWORD)
+                .toCoursesView()
+                .listMyProjects()
+                .get(0).click()
+                .toAssignmentView();
 
-		List<AssignmentsView.Assignment> assignments = view.listAssignments();
-		assertEquals(2, assignments.size());
 
-		AssignmentsView.Assignment assignment = assignments.get(0);
-		assertEquals(assignment.getName(), "Part 1. E2E & Boundary Testing");
-		assertEquals(assignment.getNumber(), "1");
-	}
+        List<AssignmentsView.Assignment> assignments = view.listAssignments();
+        assertEquals(2, assignments.size());
 
-	@Test
-	public void testAssignment() throws ParseException {
-		final AssignmentView view = openLoginScreen()
-				.login(ASSISTANT_USERNAME, ASSISTANT_PASSWORD)
-				.listAssistingCourses()
-				.get(0).click()
-				.listGroups()
-				.get(0).click()
-				.toAssignmentView()
-				.listAssignments()
-				.get(1).click();
+        AssignmentsView.Assignment assignment = assignments.get(0);
+        assertEquals(assignment.getName(), "Part 1. E2E & Boundary Testing");
+        assertEquals(assignment.getNumber(), "1");
+    }
+
+    @Test
+    public void testAssignment() throws ParseException {
+        final AssignmentView view = openLoginScreen()
+                .login(ASSISTANT_USERNAME, ASSISTANT_PASSWORD)
+                .listAssistingCourses()
+                .get(0).click()
+                .listGroups()
+                .get(0).click()
+                .toAssignmentView()
+                .listAssignments()
+                .get(1).click();
 
         final CourseEdition course = courseEditions.find(1);
         final Group group = groups.find(course).get(0);
-        final Delivery modelDelivery = deliveries.find(group,2L);
+        final Delivery modelDelivery = deliveries.find(group, 2L);
         final Assignment modelAssignment = modelDelivery.getAssignment();
         final AssignmentView.Assignment viewAssignment = view.getAssignment();
 
@@ -70,8 +72,8 @@ public class AssignmentTest extends WebTest {
         assertEquals(modelDelivery.getState(), viewAssignment.getStatus());
     }
 
-	@Test
-	public void testReviewedAssignment() throws ParseException {
+    @Test
+    public void testReviewedAssignment() throws ParseException {
         final AssignmentView view = openLoginScreen()
                 .login(ASSISTANT_USERNAME, ASSISTANT_PASSWORD)
                 .listAssistingCourses()
@@ -84,12 +86,26 @@ public class AssignmentTest extends WebTest {
 
         final CourseEdition course = courseEditions.find(1);
         final Group group = groups.find(course).get(0);
-        final Delivery modelDelivery = deliveries.find(group,1L);
+        final Delivery modelDelivery = deliveries.find(group, 1L);
         final Delivery.Review modelReview = modelDelivery.getReview();
         final AssignmentView.Review viewReview = view.getAssignment().getReview();
 
         assertEquals(modelReview.getGrade(), viewReview.getGrade().get(), 1e-4);
         assertEquals(modelReview.getReviewUser().getName(), viewReview.getReviewer().get());
         assertEquals(modelReview.getCommentary(), viewReview.getCommentary());
-	}
+    }
+
+    @Test
+    public void testAssignTAs() {
+//		final AssignmentView view =
+        LoginView login = openLoginScreen();
+        CourseView courseView = login.login(ADMIN_USERNAME, ADMIN_PASSWORD)
+                .listCourses()
+                .get(0).click();
+//                .listAssignments()
+//                .get(0).click();
+
+//                .get(0).click()
+//        .toCoursesView();
+    }
 }
